Internal workhorse join function, used in the backend of joyn

Always executes a full join.

joyn_workhorse(
  x,
  y,
  by = intersect(names(x), names(y)),
  match_type = c("1:1"),
  suffixes = getOption("joyn.suffixes")
)

Arguments

x

data object, "left" or "master"

y

data object, "right" or "using"

by

atomic character vector: key specifying join

match_type

atomic character vector of length 1: either "1:1" (default) "1:m", "m:1", or "m:m". If "m:m" then executes data.table::merge.data.table in the backend, otherwise uses collapse::join()

suffixes

atomic character vector: give suffixes to columns common to both x and y

Value

data object of same class as x

Examples

if (FALSE) {
# Full join
library(data.table)
x1 = data.table(id = c(1L, 1L, 2L, 3L, NA_integer_),
                t  = c(1L, 2L, 1L, 2L, NA_integer_),
                x  = 11:15)
y1 = data.table(id = c(1,2, 4),
                y  = c(11L, 15L, 16))
joyn:::joyn_workhorse(x = x1, y=y1)
}