style of text displayed

This is an adaptation from https://github.com/r-lib/pkgbuild/blob/3ba537ab8a6ac07d3fe11c17543677d2a0786be6/R/styles.R

Usage

style(..., sep = "")

Arguments

...

combination of type and text in the form type1 = text1, type2 = text2

sep

a character string to separate the terms to paste

Value

formatted text

See also