Merge two tables

This is the main and, basically, the only function in joyn.

Usage

merge(
  x,
  y,
  by = intersect(names(x), names(y)),
  yvars = TRUE,
  match_type = c("m:m", "m:1", "1:m", "1:1"),
  keep = c("full", "left", "master", "right", "using", "inner"),
  update_values = FALSE,
  update_NAs = update_values,
  reportvar = "report",
  reporttype = c("character", "numeric"),
  roll = NULL,
  keep_y_in_x = FALSE,
  sort = TRUE,
  verbose = getOption("joyn.verbose"),
  allow.cartesian = NULL
)

Arguments

x

data frame: referred to left in R terminology, or master in Stata terminology.

y

data frame: referred to right in R terminology, or using in Stata terminology.

by

a character vector of variables to join by. If NULL, the default, joyn will do a natural join, using all variables with common names across the two tables. A message lists the variables so that you can check they're right (to suppress the message, simply explicitly list the variables that you want to join). To join by different variables on x and y use a vector of expressions. For example, by = c("a = b", "z") will use "a" in x, "b" in y, and "z" in both tables.

yvars

character: Vector of variable names that will be kept after the merge. If TRUE (the default), it keeps all the brings all the variables in y into x. If FALSE or NULL, it does not bring any variable into x, but a report will be generated.

match_type

character: one of "m:m", "m:1", "1:m", "1:1". Default is "m:m" since this is the default generally used in joins in R. However, following Stata's recommendation, it is better to be explicit and use any of the other three match types (See details in match types sections).

keep

character: One of "full", "left", "master", "right", "using", "inner". Default is "full". Even though this is not the regular behavior of joins in R, the objective of joyn is to present a diagnosis of the join, so that it must use by default a full join. Yet, if "left" or "master", it keeps the observations that matched in both tables and the ones that did not match in x. The ones in y will be discarded. If "right" or "using", it keeps the observations that matched in both tables and the ones that did not match in y. The ones in x will be discarded. If "inner", it only keeps the observations that matched both tables.

update_values

logical: If TRUE, it will update all values of variables in x with the actual of variables in y with the same name as the ones in x. NAs from y won't be used to update actual values in x. Yet, by default, NAs in x will be updated with values in y. To avoid this, make sure to set update_NAs = FALSE

update_NAs

logical: If TRUE, it will update NA values of all variables in x with actual values of variables in y that have the same name as the ones in x. If FALSE, NA values won't be updated, even if update_values is TRUE

reportvar

character: Name of reporting variable. Default if "report". This is the same as variable "_merge" in Stata after performing a merge. If FALSE or NULL, the reporting variable will be excluded from the final table, though a summary of the join will be display after concluding.

reporttype

character: One of "character" or "numeric". Default is "character". If "numeric", the reporting variable will contain numeric codes of the source and the contents of each observation in the joined table.

roll

double: to be implemented

keep_y_in_x

logical: If TRUE, it will keep the original variable from y when both tables have common variable names. Thus, the prefix "y." will be added to the original name to distinguish from the resulting variable in the joined table.

sort

logical: If TRUE, sort by key variables in by. Default is TRUE.

verbose

logical: if FALSE, it won't display any message (programmer's option). Default is TRUE.

allow.cartesian

logical: Check documentation in official web site. Default is NULL, which implies that if the join is "1:1" it will be FALSE, but if the join has any "m" on it, it will be converted to TRUE. By specifying TRUE of FALSE you force the behavior of the join.

Value

a data.table joining x and y.

match types

Using the same wording of the Stata manual

1:1: specifies a one-to-one match merge. The variables specified in by uniquely identify single observations in both table.

1:m and m:1: specify one-to-many and many-to-one match merges, respectively. This means that in of the tables the observations are uniquely identify by the variables in by, while in the other table many (two or more) of the observations are identify by the variables in by

m:m refers to many-to-many merge. variables in by does not uniquely identify the observations in either table. Matching is performed by combining observations with equal values in by; within matching values, the first observation in the master (i.e. left or x) table is matched with the first matching observation in the using (i.e. right or y) table; the second, with the second; and so on. If there is an unequal number of observations within a group, then the last observation of the shorter group is used repeatedly to match with subsequent observations of the longer group.

Examples

# Simple merge
library(data.table)
x1 = data.table(id = c(1L, 1L, 2L, 3L, NA_integer_),
t  = c(1L, 2L, 1L, 2L, NA_integer_),
x  = 11:15)

y1 = data.table(id = 1:2,
                y  = c(11L, 15L))

x2 = data.table(id = c(1, 1, 2, 3, NA),
                t  = c(1L, 2L, 1L, 2L, NA_integer_),
                x  = c(16, 12, NA, NA, 15))

y2 = data.table(id = c(1, 2, 5, 6, 3),
              yd = c(1, 2, 5, 6, 3),
              y  = c(11L, 15L, 20L, 13L, 10L),
              x  = c(16:20))
merge(x1, y1)
#> > removing key variables `id` from yvars
#> 
#> -- JOYn Report --
#> 
#>    report n percent
#> 1:      x 2     40%
#> 2:  x & y 3     60%
#> 3:  total 5    100%
#> ---------------------------------------------------------- End of JOYn report --
#>    id  t  x  y report
#> 1:  1  1 11 11  x & y
#> 2:  1  2 12 11  x & y
#> 3:  2  1 13 15  x & y
#> 4:  3  2 14 NA      x
#> 5: NA NA 15 NA      x

# Bad merge for not specifying by argument
merge(x2, y2)
#> > removing key variables `id` and `x` from yvars
#> 
#> -- JOYn Report --
#> 
#>    report n percent
#> 1:      x 4   44.4%
#> 2:  x & y 1   11.1%
#> 3:      y 4   44.4%
#> 4:  total 9    100%
#> ---------------------------------------------------------- End of JOYn report --
#>    id  x  t yd  y report
#> 1:  1 12  2 NA NA      x
#> 2:  1 16  1  1 11  x & y
#> 3:  2 17 NA  2 15      y
#> 4:  2 NA  1 NA NA      x
#> 5:  3 20 NA  3 10      y
#> 6:  3 NA  2 NA NA      x
#> 7:  5 18 NA  5 20      y
#> 8:  6 19 NA  6 13      y
#> 9: NA 15 NA NA NA      x

# good merge, ignoring variable x from y
merge(x2, y2, by = "id")
#> > removing key variables `id` from yvars
#> i variable `x` in table y is ignored because arguments `update_NAs` and
#> `update_values` are FALSE.
#> 
#> -- JOYn Report --
#> 
#>    report n percent
#> 1:      x 1   14.3%
#> 2:  x & y 4   57.1%
#> 3:      y 2   28.6%
#> 4:  total 7    100%
#> ---------------------------------------------------------- End of JOYn report --
#>    id  t  x yd  y report
#> 1:  1  1 16  1 11  x & y
#> 2:  1  2 12  1 11  x & y
#> 3:  2  1 NA  2 15  x & y
#> 4:  3  2 NA  3 10  x & y
#> 5:  5 NA NA  5 20      y
#> 6:  6 NA NA  6 13      y
#> 7: NA NA 15 NA NA      x

# update NAs in x variable form x
merge(x2, y2, by = "id", update_NAs = TRUE)
#> > removing key variables `id` from yvars
#> 
#> -- JOYn Report --
#> 
#>        report n percent
#> 1: NA updated 2   28.6%
#> 2:          x 1   14.3%
#> 3:      x & y 2   28.6%
#> 4:          y 2   28.6%
#> 5:      total 7    100%
#> ---------------------------------------------------------- End of JOYn report --
#>    id  t  x yd  y     report
#> 1:  1  1 16  1 11      x & y
#> 2:  1  2 12  1 11      x & y
#> 3:  2  1 17  2 15 NA updated
#> 4:  3  2 20  3 10 NA updated
#> 5:  5 NA 18  5 20          y
#> 6:  6 NA 19  6 13          y
#> 7: NA NA 15 NA NA          x

# Update values in x with variables from y
merge(x2, y2, by = "id", update_values = TRUE)
#> > removing key variables `id` from yvars
#> 
#> -- JOYn Report --
#> 
#>           report n percent
#> 1:    NA updated 2   28.6%
#> 2:   not updated 1   14.3%
#> 3: value updated 1   14.3%
#> 4:         x & y 1   14.3%
#> 5:             y 2   28.6%
#> 6:         total 7    100%
#> ---------------------------------------------------------- End of JOYn report --
#>    id  t  x yd  y        report
#> 1:  1  1 16  1 11         x & y
#> 2:  1  2 16  1 11 value updated
#> 3:  2  1 17  2 15    NA updated
#> 4:  3  2 20  3 10    NA updated
#> 5:  5 NA 18  5 20             y
#> 6:  6 NA 19  6 13             y
#> 7: NA NA 15 NA NA   not updated