Merge two tables

This is the main and, basically, the only function in joyn.

merge(
  x,
  y,
  by = NULL,
  roll = NULL,
  yvars = NULL,
  join_type = c("m:m", "m:1", "1:m", "1:1"),
  keep = c("full", "left", "master", "right", "using", "inner"),
  update_values = FALSE,
  updateNA = update_values,
  reportvar = "report",
  reporttype = c("character", "numeric"),
  keep_y_in_x = FALSE,
  verbose = TRUE
)

Arguments

x
data frame: referred to left in R terminology, or master in Stata terminology.

y
data frame: referred to right in R terminology, or using in Stata terminology.

by
a character vector of variables to join by. If NULL, the default, joyn will do a natural join, using all variables with common names across the two tables. A message lists the variables so that you can check they're right (to suppress the message, simply explicitly list the variables that you want to join). To join by different variables on x and y use a vector of expressions. For example, by = c("a = b", "z") will use "a" in x, "b" in y, and "z" in both tables.

roll
double: to be implemented

yvars
character: Vector of variable names that will be kept after the merge. Be defatult it keeps all the varialbes in y into x.

join_type
character: one of "m:m", "m:1", "1:m", "1:1". Default is "m:m" since this is the default generally used in joins in R. However, following Stata's recommendation, it is better to be explicit and use any of the other three join types (See details in Join types sections).

keep
character: One of "full", "left", "master", "right", "using", "inner". Default is "full". Even though this is not the regular behavior of joins in R, the objective of joyn is to present a diagnosys of the join, so that it must use by default a full join. Yet, if "left" or "master", it keeps the observations that matched in both tables and the ones that did not match in x. The ones in y will be discarded. If "right" or "using", it keeps the observations that matched in both tables and the ones that did not match in y. The ones in x will be discarded. If "inner", it only keeps the observations that matched both tables.

update_values
logical: If TRUE, it will update all values of variables in x with the actual of variables in y with the same name as the ones in x. NAs from y won't be used to update actual values in x.

updateNA
logical: If TRUE, it will update NA values ok all variables in x with actual values of variables in y that have the same name as the ones in x. If FALSE, NA values won't be updated.

reportvar
character: Name of reporting variable. Default if "report". This is the same as variable "_merge" in Stata after performing a merge. If FALSE or NULL, the reporting variable will be excluded from the final table, though a summary of the join will be display after concluding.

reporttype
charcter: One of "character" or "numeric". Default is "character". If "numeric", the reporting variable will contain numeric codes of the source and the contents of each observation in the joined table.

keep_y_in_x
logical: If TRUE, it will keep the original variable from y when both tables have common variable names. Thus, the prefix "y." will be added to the original name to distinguish from the resulting variable in the joined table.

verbose
logical: if FALSE, it won't display any message (programmer's option). Default is TRUE.

Value

a data.table joining x and y.

Join types

Using the same wording of the Stata manual:

1:1

specifies a one-to-one match merge. The variables specified in by uniquely identify single observations in both table.

1:m and m:1

specify one-to-many and many-to-one match merges, respectively. This means that one of the tables only one observation that uniquely identifies many (i.e, two or more) observations in the other table.

m:m

many-to-many merge. variables in by does not uniquely identify the observations in either table. Matching is performed by combining observations with equal values in by; within matching values, the first observation in the master (i.e. left or x) table is matched with the first matching observation in the using (i.e. right or y) table; the second, with the second; and so on. If there is an unequal number of observations within a group, then the last observation of the shorter group is used repeatedly to match with subsequent observations of the longer group.